+++ .gitignore
@@ -1,5 +1,4 @@
-.vscode/settings.json
-
+.vscode/*
 
 # See https://help.github.com/articles/ignoring-files/ for more about ignoring files.
 
@@ -24,3 +23,4 @@
 npm-debug.log*
 yarn-debug.log*
 yarn-error.log*
+node_modules
