+++ src/components/FrontPageMain.js
