+++ src/questionData.js
@@ -1,150 +1,150 @@
 const data = [
     /* familiarity and exploration questions*/
     {
-        id: 1,
+        id: 0,
         question: "You often venture out to find new music.",
     },
     {
-        id: 2,
+        id: 0,
         question: "If a friend recommended an artist, you would actually try listening to them.",
     },
     {
-        id: 3,
+        id: 0,
         question: "It takes a while for you to get into artists you’ve never listened to before.",
     },
     {
-        id: 4,
+        id: 0,
         question: "You replay the same artists over and over again until you get sick of it.",
     },
     {
-        id: 5,
+        id: 0,
         question: "You prefer listening to mainstream music.",
     },
     {
-        id: 6,
+        id: 0,
         question: "You like listening to all the other songs your favorite artists have been featured on.",
     },
     {
-        id: 7,
+        id: 0,
         question: "You organize your playlists by artist.",
     },
     {
-        id: 8,
+        id: 0,
         question: "You research your favorite artists' discography and know a lot about them.",
     },
     {
-        id: 9,
+        id: 0,
         question: "You are a die-hard fan when it comes to your favorite artists.",
     },
     /* Loyal and variety questions*/
     {
-        id: 10,
+        id: 1,
         question: "You are a die-hard fan when it comes to your favorite songs.",
     },
     {
-        id: 11,
+        id: 1,
         question: "You repeatedly play the same set of 10 songs.",
     },
     {
-        id: 12,
+        id: 1,
         question: "You cycle through different songs every week.",
     },
     {
-        id: 13,
+        id: 1,
         question: "You have a playlist for every vibe.",
     },
     {
-        id: 14,
+        id: 1,
         question: "It takes you multiple listens to really like a song.",
     },
     {
-        id: 15,
+        id: 1,
         question: "You find certain songs invoke the same emotions and memories when you listen to them.",
     },
     {
-        id: 16,
+        id: 1,
         question: "If a friend recommended a song, you would actually listen to it.",
     },
     {
-        id: 17,
+        id: 1,
         question: "You put in time to research your favorite songs.",
     },
     {
-        id: 18,
+        id: 1,
         question: "Your friends tell you to stop playing the same song over and over again.",
     },
     /* timelessness and newness questions*/
     {
-        id: 19,
+        id: 2,
         question: "You always know when new music is releasing or has released.",
     },
     {
-        id: 20,
+        id: 2,
         question: "You prefer to listen to older music.",
     },
     {
-        id: 21,
+        id: 2,
         question: "Songs grow on you over time.",
     },
     {
-        id: 22,
+        id: 2,
         question: "You are constantly adding new songs to your playlists.",
     },
     {
-        id: 23,
+        id: 2,
         question: "You keep up to date with an artist’s activities.",
     },
     {
-        id: 24,
+        id: 2,
         question: "Throwback songs are a must have in any playlist.",
     },
     {
-        id: 25,
+        id: 2,
         question: "You are someone who is always caught up with music trends on social media apps.",
     },
     {
-        id: 26,
+        id: 2,
         question: "You miss the “old” versions of many current artists.",
     },
     {
-        id: 27,
+        id: 2,
         question: "You don’t really connect with older music vibes.",
     },
     /* commonality and uniqueness questions */
     {
-        id: 28,
+        id: 3,
         question: "You prefer small venue/bar concerts over big stadium concerts. ",
     },
     {
-        id: 29,
+        id: 3,
         question: "You actively try to find music from smaller artists.",
     },
     {
-        id: 30,
+        id: 3,
         question: "You often listen to Top 100 playlists",
     },
     {
-        id: 31,
+        id: 3,
         question: "You tend to listen to music that is currently trending on social media apps.",
     },
     {
-        id: 32,
+        id: 3,
         question: "You like the songs on the radio.",
     },
     {
-        id: 33,
+        id: 3,
         question: "You and your friends like the same type of music.",
     },
     {
-        id: 34,
+        id: 3,
         question: "You find it hard to meet people that have the same music tastes as you.",
     },
     {
-        id: 35,
+        id: 3,
         question: "You are super comfortable with having aux in the car.",
     },
     {
-        id: 36,
+        id: 3,
         question: "You want to know information about your favorite artists.",
     },
 ]
