+++ src/tools/README.md
@@ -1,2 +1,1 @@
-
 This folder contains code used to generate the user.db database
